requirements.txt
# Face-Based Image Retrieval System
# Core ML
insightface>=0.7.3
onnxruntime>=1.16.0
numpy<2

# Vector Search
faiss-cpu>=1.7.4

# Clustering
scikit-learn>=1.3.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
aiofiles>=23.0.0
watchdog>=4.0.0
